fix for sampled structure-type images were generated
Former-commit-id: 1e61f5a4222b2b056d8787fe958f3b6f2ea40bb7

+++ model_utils.py
@@ -281,7 +281,7 @@ def save_progress(enc, dec, dataProvider, logger, embedding, opt):
     
     z = list()
     if opt.nClasses > 0:
-        class_var = Variable(torch.Tensor(dataProvider.get_classes(test_inds, 'test', 'one_hot')).cuda(gpu_id), volatile=True)
+        class_var = Variable(torch.Tensor(dataProvider.get_classes(test_inds, 'test', 'one_hot').float()).cuda(gpu_id), volatile=True)
         z.append(class_var)
 
     if opt.nRef > 0:
